Finish CoinInformationBuilder Create DisplayCoins Update number objects to doubles

+++ CryptoPortfolio.Business.Contracts/CryptoBits/CoinBuy.cs
@@ -9,14 +9,14 @@ namespace CryptoPortfolio.Business.Contracts.CryptoBits
     /// </summary>
     public class CoinBuy
     {
-        public decimal quantity { get; set; }
+        public double quantity { get; set; }
         public string pair { get; set; }
-        public decimal price { get; set; }
+        public double price { get; set; }
         public DateTime buyDate { get; set; }
         public List<CoinSell> coinSellList { get; set; }
-        public decimal fee { get; set; }
+        public double fee { get; set; }
         public string feeSymbol { get; set; }
         public TransactionType transactionType { get; set; }
-        public decimal available { get; set; }
+        public double available { get; set; }
     }
 }
