+++ CryptoPortfolio.Business.Contracts/CryptoBits/CoinSell.cs
@@ -9,14 +9,14 @@ namespace CryptoPortfolio.Business.Contracts.CryptoBits
     /// </summary>
     public class CoinSell
     {
-        public decimal quantity { get; set; }
+        public double quantity { get; set; }
         public string pair { get; set; }
-        public decimal price { get; set; }
+        public double price { get; set; }
         public DateTime saleDate { get; set; }
-        public decimal fee { get; set; }
+        public double fee { get; set; }
         public string feeSymbol { get; set; }
         public TransactionType transactionType { get; set; }
-        public decimal qtySold { get; set; }
+        public double qtySold { get; set; }
         public bool processed { get; set; }
     }
 }
